Tutorial + exercises + some additional bits
This commit is the basis of an audio player, while it was created based off the tutorial all exercises have been completed independently and further functionality has been added, such as looping capabilities.

There is plans to add additional features to this application, such as gain control and zoom in and out on waveform. Additionally, it is planned to use an external library (taglib) to be able to extrapolate album artwork to draw with juce image cache. Finally, it is planned to include a list box type implementation with a vector of juce files so the user can add songs to a playlist that will play sequentially, or when chosen. A custom look and feel is also planned to some juce ;) to the application