Added more custom look and feel buttons + added rewind/fast forward
Created some more look and feel custom buttons (rewind and fastforward), also added functionality for these new buttons, being rewinding and fastforwarding. This was achieved through inheriting timer in the transport controls and using the fast forward and rewind buttons to start a timer, they also have their states added to the transport state enum. When the timercallback happens the state is checked and if it is rewind or fastforward then we +/- the position of the transport source at the rate of the timer callback.  Note: there is an issue where if the playback is paused and then you fastforward/rewind then playback does not resume when play is pressed. This is the next work to be done fixing this. Also, the switch state to handle the state is becoming large, and will have to be refactored into a state machine pattern for future development!