Basis for playlist function
Created a basis for playlist function, being a song object that takes a file and sets its member file based on what is passed to the objects constructor.

These songs are stored in a queue of uniqueptrs to the song object. While we have implemented functionality to move to next song, it is not thread safe.

Next implementation will be to make this thread safe.