Made song finished functionality thread safe
So we had made the functionality to be able to get next song from queue, however it was not thread safe, and using the timer to fix it even with 1ms callback did not always catch the stream ended to ensure that the song would switch.

This has been fixed by having the check for song finished in the processing block, and utilizing an atomic flag (bool), to ensure thread safety when quickly writing and reading from the different threads. And is now fully functional and thread safe + can have whatever timer callback speed we want now!