Added don't change song when looping enabled + hooked up prev/next buttons
Where the song is changed I just added a check to ensure that the reader source is not set to looping so the song won't change on end, unless looping is disabled.

Also hooked up next and previous button to respective functions, however this has tightly coupled the two classes maincontent and transport.

Therefore, next session should really focus on refactoring and incorporating some design patterns to loosen some of the current coupling.

Namely: state machine (for the switch case in transport control) + observer pattern for trigger next and prev song function without needing to access the buttons, publicly, from a separate class!